pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neuromorphic-ucav-sdk"
version = "0.1.0"
description = "SDK for developing biomimetic UCAVs using neuromorphic computing"
readme = "README.md"
authors = [
    {name = "Oberon Team"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "torch>=1.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12",
    "flake8>=3.9",
    "black>=21.5b2",
    "mypy>=0.812",
    "sphinx>=4.0.2",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]